WK6/project6/slackbot/requirements.txt
psycopg2-binary
requests
sqlalchemy
